Move federation api over to use OCSController - fixes #20771
